requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.3
fastapi-cli==0.0.24
fastapi-cloud-cli==0.16.0
fastar==0.9.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.2.27
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.26
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pinecone==8.1.2
pinecone-plugin-assistant==3.0.3
pinecone-plugin-interface==0.0.7
pydantic==2.12.5
pydantic-extra-types==2.11.2
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.57.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
tenacity==9.1.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
